Add fd wrapper class
Instead of returning only the textual target of an fd link,
wrap it using a class that provides more functionality.
This is a great way to prepare for smarter stitching against sockets
(for example).

Add sample code that shows how to stitch task fds against network or
unix sockets from the net API.