Return vectors instead of sets for most APIs
Except for the process enumeration API, where it feels native to return
a container sorted by the PIDs, most other APIs should probably retain
the order in the file, instead of sorting the values.

This required a change in the parser implementation, since we cannot use
`emplace` with `std::vector`. So now, instead of creating and returning
the container, we accept an inserted and let the caller decide which
container to use and how to fill it.

Added a unittest to verify that the new version of parse_lines moves the
objects.