Adjust to compilation as CMake sub-directory
**NOTE:** This change breaks backward compatibility, by changing the
CMake configuration options.

Avoid possible option collisions when using pfs as sub-directory.
Some options are supposed to be inherited from the project and are common
CMake flags, like the `CMAKE_BUILD_TYPE` or `BUILD_SHARED_LIBS`.
Whereas other flags are not supposed to be shared, and hence should are
to be prefixed with the project name.

Using `pfs_` for the prefix, rather than `PFS_` because CMake
automatically generates some library specific variables (Check that out
by looking at `CMakeCache.txt`) and uses the project name (`pfs`) as the
prefix.
This choice has been also observed in protobuf's CMake code.